Improve portal dungeon monsters. Add ability to use variables in enchantment strengths (#60788)
* Add forms and change flavor

* Compiles, untested, ugly

* Inheritance

* Update magic_enchantment.cpp

* Update magic_enchantment.cpp

* cleanup

* Add multiply

* Add new monsters

* Fixes

* Fixes

* Update nether.json

* Update magic_enchantment.cpp

* Update magic_enchantment.cpp

* Update magic_enchantment.h

* Update magic_enchantment.h

* Update magic_enchantment.cpp

* Update magic_enchantment.h

* Update dialogue.h

* Update dialogue_helpers.h
