+++ data/json/portal_storm_effect_on_condition.json
@@ -210,6 +210,8 @@
     "effect": [
       { "arithmetic": [ { "global_val": "var", "var_name": "cause_portal_storm" }, "=", { "const": 0 } ] },
       { "arithmetic": [ { "global_val": "var", "var_name": "cause_early_portal_storm" }, "=", { "const": 0 } ] },
+      { "arithmetic": [ { "global_val": "var", "var_name": "portal_dungeon_slow_strength" }, "=", { "const": 0 } ] },
+      { "arithmetic": [ { "global_val": "var", "var_name": "portal_dungeon_carry_strength" }, "=", { "const": 0 } ] },
       "next_weather",
       {
         "u_message": "As suddenly as it began, the portal storm fades.  Reality returns to normal.  Hopefully the scarring is minimal…",
@@ -1218,7 +1220,7 @@
       { "u_teleport": { "global_val": "dungeon_start" }, "fail_message": "Something is very wrong!", "force": true },
       { "arithmetic": [ { "global_val": "var", "var_name": "portal_dungeon_level" }, "=", { "const": 1 } ] },
       { "queue_eocs": "EOC_PORTAL_STORM_DUNGEON_SHIFT", "time_in_future": [ "5 seconds", "15 seconds" ] },
-      { "run_eocs": [ "EOC_PORTAL_STORM_DUNGEON_MONSTER_SPAWN", "EOC_PORTAL_STORM_DUNGEON_ENDLESS_HUNGER_SPAWN" ] },
+      { "run_eocs": [ "EOC_PORTAL_STORM_DUNGEON_REFLECTION_SPAWN" ] },
       { "u_message": "You are trapped in <global_val:place_name>!" },
       { "u_message": { "global_val": "portal_storm_voice_mood", "default_str": "none" }, "snippet": true }
     ]
@@ -1230,7 +1232,7 @@
       { "arithmetic": [ { "global_val": "var", "var_name": "portal_dungeon_level" }, "++" ] },
       { "mapgen_update": "portal_dungeon_cleanup", "target_var": { "global_val": "dungeon_loc" } },
       { "u_teleport": { "global_val": "dungeon_start" }, "fail_message": "Something is very wrong!", "force": true },
-      { "run_eocs": [ "EOC_PORTAL_STORM_DUNGEON_MONSTER_SPAWN", "EOC_PORTAL_STORM_DUNGEON_ENDLESS_HUNGER_SPAWN" ] },
+      { "run_eocs": [ "EOC_PORTAL_STORM_DUNGEON_REFLECTION_SPAWN" ] },
       { "mapgen_update": "portal_dungeon", "target_var": { "global_val": "dungeon_loc" } },
       { "u_message": "Weren't you just here? It feel thicker somehow though." },
       { "u_message": { "global_val": "portal_storm_voice_mood", "default_str": "none" }, "snippet": true }
@@ -1247,44 +1249,86 @@
   },
   {
     "type": "effect_on_condition",
-    "id": "EOC_PORTAL_STORM_DUNGEON_MONSTER_SPAWN",
-    "effect": [
-      {
-        "u_spawn_monster": "mon_shifting_mass",
-        "real_count": { "arithmetic": [ { "global_val": "var", "var_name": "portal_dungeon_level" }, "*", { "const": 2 } ], "max": 15 },
-        "hallucination_count": { "arithmetic": [ { "global_val": "var", "var_name": "portal_dungeon_level" }, "*", { "const": 2 } ], "max": 15 },
-        "min_radius": 3,
-        "max_radius": 20,
-        "spawn_message": "A creature wanders in as if the walls weren't even there.",
-        "spawn_message_plural": "Creatures wander in as if the walls weren't even there."
-      }
-    ]
-  },
-  {
-    "type": "effect_on_condition",
-    "id": "EOC_PORTAL_STORM_DUNGEON_ENDLESS_HUNGER_SPAWN",
+    "id": "EOC_PORTAL_STORM_DUNGEON_REFLECTION_SPAWN",
     "global": true,
     "condition": {
       "and": [
         { "compare_int": [ { "global_val": "var", "var_name": "portal_dungeon_level" }, ">", { "const": 0 } ] },
         {
           "compare_int": [
-            { "global_val": "monsters_nearby", "target_var": { "global_val": "dungeon_loc" }, "id": "mon_endless_hunger" },
+            { "global_val": "monsters_nearby", "target_var": { "global_val": "dungeon_loc" }, "id": "mon_pale_reflection" },
+            "<",
+            { "const": 1 }
+          ]
+        },
+        {
+          "compare_int": [
+            { "global_val": "monsters_nearby", "target_var": { "global_val": "dungeon_loc" }, "id": "mon_twisted_reflection" },
+            "<",
+            { "const": 1 }
+          ]
+        },
+        {
+          "compare_int": [
+            { "global_val": "monsters_nearby", "target_var": { "global_val": "dungeon_loc" }, "id": "mon_dark_reflection" },
+            "<",
+            { "const": 1 }
+          ]
+        },
+        {
+          "compare_int": [
+            { "global_val": "monsters_nearby", "target_var": { "global_val": "dungeon_loc" }, "id": "mon_better_half" },
             "<",
             { "const": 1 }
           ]
         }
       ]
     },
-    "effect": [
-      {
-        "u_spawn_monster": "mon_endless_hunger",
-        "real_count": 1,
-        "min_radius": 8,
-        "max_radius": 12,
-        "spawn_message": "Something your mind can't fully parse appears, despite the distance and obstacles you know where it is.  You can feel its presence as if there were a bond of some sort between you."
-      }
-    ]
+    "effect": {
+      "switch": { "global_val": "var", "var_name": "portal_dungeon_reflect_strength" },
+      "cases": [
+        {
+          "case": 0,
+          "effect": {
+            "u_spawn_monster": "mon_pale_reflection",
+            "real_count": 1,
+            "min_radius": 8,
+            "max_radius": 12,
+            "spawn_message": "Something your mind can't fully parse appears, despite the distance and obstacles you know where it is.  You can feel its presence as if it was a part of you."
+          }
+        },
+        {
+          "case": 5,
+          "effect": {
+            "u_spawn_monster": "mon_twisted_reflection",
+            "real_count": 1,
+            "min_radius": 8,
+            "max_radius": 12,
+            "spawn_message": "Something your mind can't fully parse appears, despite the distance and obstacles you know where it is.  You can feel its presence as if it was a part of you."
+          }
+        },
+        {
+          "case": 10,
+          "effect": {
+            "u_spawn_monster": "mon_dark_reflection",
+            "real_count": 1,
+            "min_radius": 8,
+            "max_radius": 12,
+            "spawn_message": "Something your mind can't fully parse appears, despite the distance and obstacles you know where it is.  You can feel its presence as if it was a part of you."
+          }
+        },
+        {
+          "case": 15,
+          "effect": {
+            "u_spawn_monster": "mon_better_half",
+            "real_count": 1,
+            "min_radius": 8,
+            "max_radius": 12,
+            "spawn_message": "Something your mind can't fully parse appears, despite the distance and obstacles you know where it is.  You can feel its presence as if it was a part of you."
+          }
+        }
+      ]
+    }
   },
   {
     "type": "effect_on_condition",
