+++ src/magic_enchantment.cpp
@@ -4,11 +4,13 @@
 #include <set>
 #include <string>
 
+#include "avatar.h"
 #include "calendar.h"
 #include "character.h"
 #include "condition.h"
 #include "creature.h"
 #include "debug.h"
+#include "dialogue_helpers.h"
 #include "enum_conversions.h"
 #include "enums.h"
 #include "generic_factory.h"
@@ -268,7 +270,7 @@ void enchantment::bodypart_changes::serialize( JsonOut &jsout ) const
 }
 
 void enchantment::load( const JsonObject &jo, const std::string &,
-                        const cata::optional<std::string> &inline_id )
+                        const cata::optional<std::string> &inline_id, bool is_child )
 {
     optional( jo, was_loaded, "id", id, enchantment_id( inline_id.value_or( "" ) ) );
 
@@ -300,7 +302,7 @@ void enchantment::load( const JsonObject &jo, const std::string &,
     if( jo.has_string( "condition" ) ) {
         std::string condit;
         optional( jo, was_loaded, "condition", condit );
-        cata::optional<enchantment::condition> con = io::string_to_enum_optional<condition>( condit );
+        cata::optional<condition> con = io::string_to_enum_optional<condition>( condit );
         if( con.has_value() ) {
             active_conditions.second = con.value();
         } else {
@@ -320,11 +322,37 @@ void enchantment::load( const JsonObject &jo, const std::string &,
     optional( jo, was_loaded, "modified_bodyparts", modified_bodyparts );
     optional( jo, was_loaded, "mutations", mutations );
 
+    if( !is_child && jo.has_array( "values" ) ) {
+        for( const JsonObject value_obj : jo.get_array( "values" ) ) {
+            const enchant_vals::mod value = io::string_to_enum<enchant_vals::mod>
+                                            ( value_obj.get_string( "value" ) );
+            if( value_obj.has_member( "add" ) ) {
+                int_or_var<dialogue> add = get_int_or_var<dialogue>( value_obj, "add", false );
+                values_add.emplace( value, add );
+            }
+            if( value_obj.has_member( "multiply" ) ) {
+                int_or_var<dialogue> mult;
+                if( value_obj.has_float( "multiply" ) ) {
+                    mult.max.int_val = mult.min.int_val = value_obj.get_float( "multiply" ) * 100;
+                } else {
+                    mult = get_int_or_var<dialogue>( value_obj, "multiply", false );
+
+                }
+                values_multiply.emplace( value, mult );
+            }
+        }
+    }
+}
+
+void enchant_cache::load( const JsonObject &jo, const std::string &,
+                          const cata::optional<std::string> &inline_id )
+{
+    enchantment::load( jo, "", inline_id, true );
     if( jo.has_array( "values" ) ) {
         for( const JsonObject value_obj : jo.get_array( "values" ) ) {
             const enchant_vals::mod value = io::string_to_enum<enchant_vals::mod>
                                             ( value_obj.get_string( "value" ) );
-            const int add = value_obj.get_int( "add", 0 );
+            const double add = value_obj.get_int( "add", 0 );
             const double mult = value_obj.get_float( "multiply", 0.0 );
             if( add != 0 ) {
                 values_add.emplace( value, add );
@@ -336,19 +364,12 @@ void enchantment::load( const JsonObject &jo, const std::string &,
     }
 }
 
-void enchantment::serialize( JsonOut &jsout ) const
+void enchant_cache::serialize( JsonOut &jsout ) const
 {
     jsout.start_object();
 
-    if( !id.is_empty() ) {
-        jsout.member( "id", id );
-        jsout.end_object();
-        // if the enchantment has an id then it is defined elsewhere and does not need to be serialized.
-        return;
-    }
-
-    jsout.member( "has", io::enum_to_string<has>( active_conditions.first ) );
-    jsout.member( "condition", io::enum_to_string<condition>( active_conditions.second ) );
+    jsout.member( "has", io::enum_to_string<enchantment::has>( active_conditions.first ) );
+    jsout.member( "condition", io::enum_to_string<enchantment::condition>( active_conditions.second ) );
     if( emitter ) {
         jsout.member( "emitter", emitter );
     }
@@ -395,9 +416,6 @@ void enchantment::serialize( JsonOut &jsout ) const
     jsout.start_array();
     for( int value = 0; value < static_cast<int>( enchant_vals::mod::NUM_MOD ); value++ ) {
         enchant_vals::mod enum_value = static_cast<enchant_vals::mod>( value );
-        if( get_value_add( enum_value ) == 0 && get_value_multiply( enum_value ) == 0.0 ) {
-            continue;
-        }
         jsout.start_object();
         jsout.member( "value", io::enum_to_string<enchant_vals::mod>( enum_value ) );
         if( get_value_add( enum_value ) != 0 ) {
@@ -413,12 +431,21 @@ void enchantment::serialize( JsonOut &jsout ) const
     jsout.end_object();
 }
 
-bool enchantment::stacks_with( const enchantment &rhs ) const
+bool enchant_cache::stacks_with( const enchantment &rhs ) const
 {
     return active_conditions == rhs.active_conditions;
 }
 
-bool enchantment::add( const enchantment &rhs )
+bool enchant_cache::add( const enchantment &rhs, Character &you )
+{
+    if( !stacks_with( rhs ) ) {
+        return false;
+    }
+    force_add( rhs, you );
+    return true;
+}
+
+bool enchant_cache::add( const enchant_cache &rhs )
 {
     if( !stacks_with( rhs ) ) {
         return false;
@@ -427,9 +454,10 @@ bool enchantment::add( const enchantment &rhs )
     return true;
 }
 
-void enchantment::force_add( const enchantment &rhs )
+void enchant_cache::force_add( const enchant_cache &rhs )
 {
-    for( const std::pair<const enchant_vals::mod, int> &pair_values : rhs.values_add ) {
+    for( const std::pair<const enchant_vals::mod, int> &pair_values :
+         rhs.values_add ) {
         values_add[pair_values.first] += pair_values.second;
     }
     for( const std::pair<const enchant_vals::mod, double> &pair_values : rhs.values_multiply ) {
@@ -464,32 +492,92 @@ void enchantment::force_add( const enchantment &rhs )
     }
 }
 
-void enchantment::set_has( enchantment::has value )
+void enchant_cache::force_add( const enchantment &rhs, const Character &guy )
+{
+    dialogue d( get_talker_for( guy ), nullptr );
+    for( const std::pair<const enchant_vals::mod, int_or_var<dialogue>> &pair_values :
+         rhs.values_add ) {
+        values_add[pair_values.first] += pair_values.second.evaluate( d );
+    }
+    for( const std::pair<const enchant_vals::mod, int_or_var<dialogue>> &pair_values :
+         rhs.values_multiply ) {
+        // values do not multiply against each other, they add.
+        // so +10% and -10% will add to 0%
+        values_multiply[pair_values.first] += 0.01 * pair_values.second.evaluate( d );
+    }
+
+    hit_me_effect.insert( hit_me_effect.end(), rhs.hit_me_effect.begin(), rhs.hit_me_effect.end() );
+
+    hit_you_effect.insert( hit_you_effect.end(), rhs.hit_you_effect.begin(), rhs.hit_you_effect.end() );
+
+    ench_effects.insert( rhs.ench_effects.begin(), rhs.ench_effects.end() );
+
+    if( rhs.emitter ) {
+        emitter = rhs.emitter;
+    }
+
+    for( const bodypart_changes &bp : rhs.modified_bodyparts ) {
+        modified_bodyparts.emplace_back( bp );
+    }
+
+    for( const trait_id &branch : rhs.mutations ) {
+        mutations.emplace( branch );
+    }
+
+    for( const std::pair<const time_duration, std::vector<fake_spell>> &act_pair :
+         rhs.intermittent_activation ) {
+        for( const fake_spell &fake : act_pair.second ) {
+            intermittent_activation[act_pair.first].emplace_back( fake );
+        }
+    }
+}
+
+void enchant_cache::set_has( enchantment::has value )
 {
     active_conditions.first = value;
 }
 
-void enchantment::add_value_add( enchant_vals::mod value, int add_value )
+void enchant_cache::add_value_add( enchant_vals::mod value, int add_value )
 {
     values_add[value] = add_value;
 }
 
-void enchantment::add_value_mult( enchant_vals::mod value, float mult_value )
+void enchant_cache::add_value_mult( enchant_vals::mod value, float mult_value )
 {
     values_multiply[value] = mult_value;
 }
 
-void enchantment::add_hit_me( const fake_spell &sp )
+void enchant_cache::add_hit_me( const fake_spell &sp )
 {
     hit_me_effect.push_back( sp );
 }
 
-void enchantment::add_hit_you( const fake_spell &sp )
+void enchant_cache::add_hit_you( const fake_spell &sp )
 {
     hit_you_effect.push_back( sp );
 }
 
-int enchantment::get_value_add( const enchant_vals::mod value ) const
+int enchantment::get_value_add( const enchant_vals::mod value, const Character &guy ) const
+{
+    const auto found = values_add.find( value );
+    if( found == values_add.cend() ) {
+        return 0;
+    }
+    dialogue d( get_talker_for( guy ), nullptr );
+    return found->second.evaluate( d );
+}
+
+double enchantment::get_value_multiply( const enchant_vals::mod value, const Character &guy ) const
+{
+    const auto found = values_multiply.find( value );
+    if( found == values_multiply.cend() ) {
+        return 0;
+    }
+    dialogue d( get_talker_for( guy ), nullptr );
+    return found->second.evaluate( d ) * 0.01;
+}
+
+int enchant_cache::get_value_add( const enchant_vals::mod value ) const
 {
     const auto found = values_add.find( value );
     if( found == values_add.cend() ) {
@@ -498,7 +586,7 @@ int enchantment::get_value_add( const enchant_vals::mod value ) const
     return found->second;
 }
 
-double enchantment::get_value_multiply( const enchant_vals::mod value ) const
+double enchant_cache::get_value_multiply( const enchant_vals::mod value ) const
 {
     const auto found = values_multiply.find( value );
     if( found == values_multiply.cend() ) {
@@ -507,14 +595,14 @@ double enchantment::get_value_multiply( const enchant_vals::mod value ) const
     return found->second;
 }
 
-double enchantment::modify_value( const enchant_vals::mod mod_val, double value ) const
+double enchant_cache::modify_value( const enchant_vals::mod mod_val, double value ) const
 {
     value += get_value_add( mod_val );
     value *= 1.0 + get_value_multiply( mod_val );
     return value;
 }
 
-units::energy enchantment::modify_value( const enchant_vals::mod mod_val,
+units::energy enchant_cache::modify_value( const enchant_vals::mod mod_val,
         units::energy value ) const
 {
     value += units::from_millijoule<int>( get_value_add( mod_val ) );
@@ -522,15 +610,15 @@ units::energy enchantment::modify_value( const enchant_vals::mod mod_val,
     return value;
 }
 
-units::mass enchantment::modify_value( const enchant_vals::mod mod_val,
-                                       units::mass value ) const
+units::mass enchant_cache::modify_value( const enchant_vals::mod mod_val,
+        units::mass value ) const
 {
     value += units::from_gram<int>( get_value_add( mod_val ) );
     value *= 1.0 + get_value_multiply( mod_val );
     return value;
 }
 
-int enchantment::mult_bonus( enchant_vals::mod value_type, int base_value ) const
+int enchant_cache::mult_bonus( enchant_vals::mod value_type, int base_value ) const
 {
     return get_value_multiply( value_type ) * base_value;
 }
@@ -543,7 +631,7 @@ bool enchantment::modifies_bodyparts() const
 body_part_set enchantment::modify_bodyparts( const body_part_set &unmodified ) const
 {
     body_part_set modified( unmodified );
-    for( const enchantment::bodypart_changes &changes : modified_bodyparts ) {
+    for( const bodypart_changes &changes : modified_bodyparts ) {
         if( !changes.gain.is_empty() ) {
             modified.set( changes.gain );
         }
@@ -554,7 +642,7 @@ body_part_set enchantment::modify_bodyparts( const body_part_set &unmodified ) c
     return modified;
 }
 
-void enchantment::activate_passive( Character &guy ) const
+void enchant_cache::activate_passive( Character &guy ) const
 {
     guy.mod_str_bonus( get_value_add( enchant_vals::mod::STRENGTH ) );
     guy.mod_str_bonus( mult_bonus( enchant_vals::mod::STRENGTH, guy.get_str_base() ) );
@@ -588,21 +676,21 @@ void enchantment::activate_passive( Character &guy ) const
     }
 }
 
-void enchantment::cast_hit_you( Character &caster, const Creature &target ) const
+void enchant_cache::cast_hit_you( Character &caster, const Creature &target ) const
 {
     for( const fake_spell &sp : hit_you_effect ) {
         cast_enchantment_spell( caster, &target, sp );
     }
 }
 
-void enchantment::cast_hit_me( Character &caster, const Creature *target ) const
+void enchant_cache::cast_hit_me( Character &caster, const Creature *target ) const
 {
     for( const fake_spell &sp : hit_me_effect ) {
         cast_enchantment_spell( caster, target, sp );
     }
 }
 
-void enchantment::cast_enchantment_spell( Character &caster, const Creature *target,
+void enchant_cache::cast_enchantment_spell( Character &caster, const Creature *target,
         const fake_spell &sp ) const
 {
     // check the chances
@@ -633,10 +721,30 @@ void enchantment::cast_enchantment_spell( Character &caster, const Creature *tar
     }
 }
 
-bool enchantment::operator==( const enchantment &rhs ) const
+bool enchant_cache::operator==( const enchant_cache &rhs ) const
 {
+    if( this->values_add.size() != rhs.values_add.size() ||
+        this->values_multiply.size() != rhs.values_multiply.size() ) {
+        return false;
+    }
+    auto iter_add = this->values_add.cbegin();
+    auto iter_add2 = rhs.values_add.cbegin();
+    while( iter_add != this->values_add.cend() && iter_add2 != rhs.values_add.cend() ) {
+        if( iter_add->second != iter_add2->second || iter_add->first != iter_add2->first ) {
+            return false;
+        }
+        iter_add++;
+        iter_add2++;
+    }
+    auto iter_mult = this->values_multiply.cbegin();
+    auto iter_mult2 = rhs.values_multiply.cbegin();
+    while( iter_mult != this->values_multiply.cend() && iter_mult2 != rhs.values_multiply.cend() ) {
+        if( iter_mult->second != iter_mult2->second || iter_mult->first != iter_mult2->first ) {
+            return false;
+        }
+        iter_mult++;
+        iter_mult2++;
+    }
     return this->id == rhs.id &&
-           this->get_mutations() == rhs.get_mutations() &&
-           this->values_multiply == rhs.values_multiply &&
-           this->values_add == rhs.values_add;
+           this->get_mutations() == rhs.get_mutations();
 }
